+++ src/xkcdSearchHandler.ts
